refactor: 提取库存管理组件
